feat: Add resource tracking methods
refs: RUMM-3409

+++ packages/Datadog.Unity/Runtime/Rum/DdRum.cs
@@ -66,7 +66,7 @@ namespace Datadog.Unity.Rum
         public void StartResourceLoading(string key, RumHttpMethod httpMethod, string url,
             Dictionary<string, object> attributes = null);
 
-        public void StopResourceLoading(string key, RumResourceType kind, int? statusCode = null, int? size = null,
+        public void StopResourceLoading(string key, RumResourceType kind, int? statusCode = null, long? size = null,
             Dictionary<string, object> attributes = null);
 
         public void StopResourceLoading(string key, Exception error, Dictionary<string, object> attributes = null);
@@ -118,7 +118,7 @@ namespace Datadog.Unity.Rum
             throw new NotImplementedException();
         }
 
-        public void StopResourceLoading(string key, RumResourceType kind, int? statusCode = null, int? size = null,
+        public void StopResourceLoading(string key, RumResourceType kind, int? statusCode = null, long? size = null,
             Dictionary<string, object> attributes = null)
         {
         }
